requirements.txt
packaging
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.16
pillow==10.1.0
numpy==1.26.3
scikit-learn==1.3.2
faiss-cpu==1.8.0
python-multipart==0.0.6
pydantic==2.6.1
pydantic-settings==2.1.0
setuptools>=68.0
openai==2.6.0